Updates to exportSelChan,flattenSelChan,ConvertToPaintable
Export Selected Channels + Duplicate Flatten Channels
-----------------------------------
The currently active channel will be pre-selected in the interface

Objects + Channels are now sorted alphabetically to be in line with
Mari's Object and Channels Palettes.

Only channels from the current object will be displayed by default now.
A new checkbox 'List all Objects' was added  to return to the previous
behavior.
The current object's channels will be bumped to the top of the list.

Channels (with cryptic names) created by layeredShader stacks will no
longer appear in the list

Convert to paintable
-----------------------------------
The interface was removed. Since it is so heavily used it is easier and
less time consuming for users to just make it work off the selection

Some additional logic was built in to make it work with floating/pinned
and docked channels that are not currently the active channel but are
accessible for the user in the interface